Fix incorrect logic with reference checks
Fields can reference more than one reference list. The logic previously
threw an error when the field wasn't a part of any of the lists.
Obviously in retrospect, this means that an error would be thrown for
any field that had more than one reference list specified. The logic was
inverted such that an error is thrown if the field is not in at least
one of the reference lists specified.

This also fixes the 'OrList' extension for the case of one item in the
list.